requirements.txt
Python==3.11.1
django==4.1.4
requests==2.28.2
djangorestframework==3.14.0
requests-oauthlib==1.3.1
djangorestframework-simplejwt==5.2.2

#Install these packages using pip